Fix new exceptions found by mypy
 * Remove implicit Optional types
 * Correct type for parameters in distance
 * Update system check from platform.system() to sys.platform. This is
   recognized by mypy
 * Add and fix types in HintObjects

+++ src/acconeer/exptool/a121/algo/distance/_processors.py
@@ -570,8 +570,8 @@ class Processor(ProcessorBase[ProcessorConfig, ProcessorResult]):
     @staticmethod
     def _calculate_cfar_threshold(
         abs_sweep: npt.NDArray[np.float_],
-        window_length: float,
-        guard_half_length: float,
+        window_length: int,
+        guard_half_length: int,
         num_stds: float,
         abs_noise_std: npt.NDArray,
     ) -> npt.NDArray[np.float_]:
